04-Notepad++/requirements.txt
streamlit
reportlab
pandas
openpyxl
numpy
pyautogui

pytesseract
Pillow
opencv-python
python-docx
